requirements.txt
horovod~=0.16.0
tensorflow-gpu~=1.14.0
keras==2.2.5
tensorpack>=0.9.6
git+https://www.github.com/keras-team/keras-contrib.git
msgpack
numpy
opencv-python
h5py
tqdm
image
Pillow
dotmap
python-telegram-bot
python-prctl
matplotlib==3.0.3
comet_ml
cv2